[#3] Add support for multiple tilesheets
- Change map-editor.lua to use a directory containing tilesheets. From
now on it will assume that each PNG file (tilsheet) in this directory
corresponds 1:1 to a tile type! I think it is best to keep the source
tilesheets as granular as possible - this should help avoid merge
conflicts when there are multiple artists. This is only theoretical as I
do not expect to be able to join a team of artists anytime soon but
foresight is good. Note that the individual spritesheets and tilesheets
**will** be compiled into one big spritesheet eventually, although that
is not implemented yet (see [#8]).

- Fix NativeFile#read @ native-file.lua, which used the wrong mode
(`"r"` instead of `"rb"`) and thus could not open binary files, e.g. PNG

- Fix NativeFile#getFiles and NativeFile#getDirectories @
native-file.lua, which had a dot (`.`) instead of a colon (`:`) when
calling its #isDirectory method - this is why proper type checking is
necessary